Use start time in data filenames
We had been using the timestamp in the "next" url, expecting that it would change from one page of
results to the next.  However, the providers with end_times in the pagination urls used the same
end_time throughout.

Since not all providers even do this, we now just use the inital start time as the filename for all
pages of response data.

Updates #5